ShopListViewModelのなかで、Mapの最後の位置・ズームレベルを記録していたのを別のViewModelに切り出す。
